+++ rswag-specs/lib/rswag/specs/example_group_helpers.rb
@@ -126,24 +126,13 @@ module Rswag
 
         description ||= "returns a #{metadata[:response][:code]} response"
 
-        if RSPEC_VERSION < 3
-          before do
-            submit_request(example.metadata)
-          end
+        before do |example|
+          submit_request(example.metadata)
+        end
 
-          it description, *args, **options do
-            assert_response_matches_metadata(metadata)
-            block.call(response) if block_given?
-          end
-        else
-          before do |example|
-            submit_request(example.metadata)
-          end
-
-          it description, *args, **options do |example|
-            assert_response_matches_metadata(example.metadata, &block)
-            example.instance_exec(response, &block) if block_given?
-          end
+        it description, *args, **options do |example|
+          assert_response_matches_metadata(example.metadata, &block)
+          example.instance_exec(response, &block) if block_given?
         end
       end
     end