+++ rswag-specs/lib/rswag/specs/response_validator.rb
@@ -13,11 +13,11 @@ module Rswag
       end
 
       def validate!(metadata, response)
-        swagger_doc = @config.get_openapi_spec(metadata[:openapi_spec] || metadata[:swagger_doc])
+        openapi_spec = @config.get_openapi_spec(metadata[:openapi_spec])
 
         validate_code!(metadata, response)
         validate_headers!(metadata, response.headers)
-        validate_body!(metadata, swagger_doc, response.body)
+        validate_body!(metadata, openapi_spec, response.body)
       end
 
       private
@@ -41,7 +41,7 @@ module Rswag
           is_nullable = nullable_attribute.nil? ? false : nullable_attribute
           is_required = required_attribute.nil? ? true : required_attribute
 
-          if headers.exclude?(name.to_s) && is_required
+          if !headers.include?(name.to_s) && is_required
             raise UnexpectedResponse, "Expected response header #{name} to be present"
           end
 
@@ -51,12 +51,11 @@ module Rswag
         end
       end
 
-      def validate_body!(metadata, swagger_doc, body)
+      def validate_body!(metadata, openapi_spec, body)
         response_schema = metadata[:response][:schema]
         return if response_schema.nil?
 
-        version = @config.get_openapi_spec_version(metadata[:openapi_spec] || metadata[:swagger_doc])
-        schemas = definitions_or_component_schemas(swagger_doc, version)
+        schemas = { components: openapi_spec[:components] }
 
         validation_schema = response_schema
           .merge('$schema' => 'http://tempuri.org/rswag/specs/extended_schema')
@@ -73,38 +72,13 @@ module Rswag
       end
 
       def validation_options_from(metadata)
-        is_strict = @config.openapi_strict_schema_validation
-
-        if metadata.key?(:swagger_strict_schema_validation)
-          Rswag::Specs.deprecator.warn('Rswag::Specs: WARNING: This option will be removed in v3.0 please use openapi_all_properties_required and openapi_no_additional_properties set to true')
-          is_strict = !!metadata[:swagger_strict_schema_validation]
-        elsif metadata.key?(:openapi_strict_schema_validation)
-          Rswag::Specs.deprecator.warn('Rswag::Specs: WARNING: This option will be removed in v3.0 please use openapi_all_properties_required and openapi_no_additional_properties set to true')
-          is_strict = !!metadata[:openapi_strict_schema_validation]
-        end
-
         all_properties_required = metadata.fetch(:openapi_all_properties_required, @config.openapi_all_properties_required)
         no_additional_properties = metadata.fetch(:openapi_no_additional_properties, @config.openapi_no_additional_properties)
 
         {
-          strict: is_strict,
           allPropertiesRequired: all_properties_required,
           noAdditionalProperties: no_additional_properties
         }
-      end
-
-      def definitions_or_component_schemas(swagger_doc, version)
-        if version.start_with?('2')
-          swagger_doc.slice(:definitions)
-        else # Openapi3
-          if swagger_doc.key?(:definitions)
-            Rswag::Specs.deprecator.warn('Rswag::Specs: WARNING: definitions is replaced in OpenAPI3! Rename to components/schemas (in swagger_helper.rb)')
-            swagger_doc.slice(:definitions)
-          else
-            components = swagger_doc[:components] || {}
-            { components: components }
-          end
-        end
       end
     end