+++ test-app/spec/integration/blogs_spec.rb
@@ -1,11 +1,7 @@
-require 'swagger_helper'
+require 'openapi_helper'
 
-RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
-  let(:api_key) { 'fake_key' }
-
-  before do
-    # allow(Rswag::Specs.deprecator).to receive(:warn) # Silence deprecation output from specs
-  end
+RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/openapi.json' do
+  let(:blog) { Blog.create(title: 'foo', content: 'bar', thumbnail: 'thumbnail.png') }
 
   path '/blogs' do
     post 'Creates a blog' do
@@ -14,19 +10,22 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       operationId 'createBlog'
       consumes 'application/json'
       produces 'application/json'
-      parameter name: :blog, in: :body, schema: { '$ref' => '#/components/schemas/blog' }
 
-      let(:blog) { { title: 'foo', content: 'bar', status: 'published' } }
+      let(:request_params) { { "blog" => { title: 'foo', content: 'bar', status: 'published' } } }
+      parameter name: :blog, in: :body, schema: {
+        type: :object,
+        required: true
+      }
 
       response '201', 'blog created' do
-        # schema '$ref' => '#/definitions/blog'
+        schema '$ref' => '#/components/schemas/blog'
         run_test!
       end
 
       response "422", "invalid request" do
-        schema "$ref" => "#/components/schemas/errors_object"
+        schema '$ref' => '#/components/schemas/errors_object'
 
-        let(:blog) { {title: "foo"} }
+        let(:request_params) { { "blog" => { title: 'foo' } } }
 
         run_test!
 
@@ -42,17 +41,25 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       description 'Searches blogs by keywords'
       operationId 'searchBlogs'
       produces 'application/json'
-      parameter name: :keywords, in: :query, type: 'string'
-      parameter name: :status, in: :query, getter: :blog_status,
-                enum: { 'draft': 'Retrieves draft blogs', 'published': 'Retrieves published blogs', 'archived': 'Retrieves archived blogs' },
+      parameter name: "keywords", in: :query, schema: { type: 'string' }
+      parameter name: "status",
+                in: :query,
+                schema: {
+                  type: :string,
+                  items:  {
+                    enum: { 'draft': 'Retrieves draft blogs', 'published': 'Retrieves published blogs', 'archived': 'Retrieves archived blogs' },
+                  }
+                },
                 description: 'Filter by status'
 
       before do
         Blog.create(title: 'foo', content: 'hello world', status: :published)
       end
 
-      let(:keywords) { 'foo bar' }
-      let(:blog_status) { 'published' }
+      let(:request_params) { {
+        "keywords" => 'foo bar',
+        "status" => 'published'
+       } }
 
       response '200', 'success' do
         schema type: 'array', items: { '$ref' => '#/components/schemas/blog' }
@@ -65,7 +72,9 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       response '200', 'no content' do
         schema type: 'array', items: { '$ref' => '#/components/schemas/blog' }
 
-        let(:blog_status) { 'invalid' }
+        let(:request_params) { {
+          "status" => 'invalid'
+         } }
 
         run_test! do
           expect(JSON.parse(response.body).size).to eq(0)
@@ -73,7 +82,7 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       end
 
       response '406', 'unsupported accept header' do
-        let(:'Accept') { 'application/foo' }
+        let(:request_headers) { { 'Accept' => 'application/foo' } }
         run_test!
       end
     end
@@ -87,27 +96,31 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       consumes 'application/json'
       produces 'application/json'
 
-      parameter name: :flexible_blog, in: :body, schema: {
+      parameter name: "flexible_blog", in: :body, schema: {
         oneOf: [
-          { '$ref' => '#/definitions/blog' },
-          { '$ref' => '#/definitions/flexible_blog' }
+          { '$ref' => '#/components/schemas/blog' },
+          { '$ref' => '#/components/schemas/flexible_blog' }
         ]
       }
 
-      let(:flexible_blog) { { blog: { headline: 'my headline', text: 'my text' } } }
+      let(:request_params) { { "flexible_blog" => { blog: { headline: 'my headline', text: 'my text' } } } }
 
       response '201', 'flexible blog created' do
-        schema oneOf: [{ '$ref' => '#/components/schemas/blog' }, { '$ref' => '#/components/schemas/flexible_blog' }]
+        schema oneOf: [
+          { '$ref' => '#/components/schemas/blog' },
+          { '$ref' => '#/components/schemas/flexible_blog' }
+        ]
         run_test!
       end
     end
   end
 
   path '/blogs/{id}' do
-    parameter name: :id, in: :path, type: :string
+    parameter name: 'id', in: :path, schema: { type: :string }
 
-    let(:id) { blog.id }
-    let(:blog) { Blog.create(title: 'foo', content: 'bar', thumbnail: 'thumbnail.png') }
+    let(:request_params) { {
+      "id" => blog.id
+    } }
 
     get 'Retrieves a blog' do
       tags 'Blogs'
@@ -144,13 +157,7 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
           thumbnail: 'thumbnail.png'
         }
 
-        let(:id) { blog.id }
-
         run_test!
-
-        context 'when openapi_strict_schema_validation is true' do
-          run_test!(openapi_strict_schema_validation: true)
-        end
 
         context 'when openapi_all_properties_required is true' do
           run_test!(openapi_all_properties_required: true)
@@ -162,15 +169,7 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       end
 
       response '404', 'blog not found' do
-        let(:id) { 'invalid' }
-        run_test!
-      end
-
-      response '200', 'blog found - openapi_strict_schema_validation = true', openapi_strict_schema_validation: true do
-        schema '$ref' => '#/components/schemas/blog'
-
-        let(:id) { blog.id }
-
+        let(:request_params) { { "id" => 'invalid' } }
         run_test!
       end
 
@@ -193,10 +192,14 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
   end
 
   path '/blogs/{id}/upload' do
-    parameter name: :id, in: :path, type: :string
+    parameter name: "id", in: :path, schema: { type: :string }
 
-    let(:id) { blog.id }
     let(:blog) { Blog.create(title: 'foo', content: 'bar') }
+    let(:request_params) { {
+      "id" => blog.id,
+      "blog" => blog,
+      "file" => file
+    } }
 
     put 'Uploads a blog thumbnail' do
       tags 'Blogs'
@@ -204,10 +207,10 @@ RSpec.describe 'Blogs API', type: :request, openapi_spec: 'v1/swagger.json' do
       operationId 'uploadThumbnailBlog'
       consumes 'multipart/form-data'
       parameter(
-        name: :file,
+        name: "file",
         description: "The content of the blog thumbnail",
         in: :formData,
-        type: :file,
+        schema: { type: :file },
         required: true
       )